+++ src/main/java/roomescape/controller/AdminController.java
@@ -1,0 +1,20 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@RequestMapping("/admin")
+@Controller
+public class AdminController {
+
+    @GetMapping
+    public String admin() {
+        return "admin/index";
+    }
+
+    @GetMapping("/reservation")
+    public String reservation() {
+        return "admin/reservation-legacy";
+    }
+}
